# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gigs API",
        "description": "API documentation for Gigs API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "POST register user",
                "description": "Returns user data",
                "operationId": "RegisterUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "POST Login user",
                "description": "Returns user data",
                "operationId": "LoginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserLogin"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Invalid Credentials"
                    },
                    "422": {
                        "description": "Invalid request body"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "POST Logout user",
                "description": "Logout user",
                "operationId": "Logout",
                "responses": {
                    "204": {
                        "description": "Successfully logged out"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "User's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "User's last name",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User's email address",
                        "type": "string",
                        "example": "john@doe.com"
                    },
                    "posted_rate": {
                        "title": "Posted Rate",
                        "description": "User's posted rate",
                        "type": "number",
                        "format": "double",
                        "example": 4.5
                    },
                    "companies": {
                        "title": "user companies",
                        "description": "user companies",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "company name"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "company description"
                                },
                                "address": {
                                    "type": "string",
                                    "example": "company address"
                                },
                                "number_of_posted_gigs": {
                                    "type": "integer",
                                    "example": 45
                                },
                                "number_of_started_gigs": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "user": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "johndoe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "gigs": {
                        "title": "user gigs",
                        "description": "user gigs",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "gig name"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "gig description"
                                },
                                "timestamp_start": {
                                    "type": "string",
                                    "example": "09/24/2024 12:00 PM"
                                },
                                "timestamp_end": {
                                    "type": "string",
                                    "example": "09/25/2024 12:00 PM"
                                },
                                "number_of_positions": {
                                    "type": "integer",
                                    "example": 45
                                },
                                "pay_per_hour": {
                                    "type": "number",
                                    "format": "double",
                                    "example": 4.5
                                },
                                "status": {
                                    "type": "boolean",
                                    "example": 0
                                },
                                "company": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "company_name"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "UserLogin": {
                "title": "User Login",
                "description": "The user login return by the response",
                "properties": {
                    "token_type": {
                        "title": "Token type",
                        "description": "This will always be 'Bearer'",
                        "type": "string",
                        "example": "Bearer"
                    },
                    "expires_in": {
                        "title": "Expires In",
                        "description": "Expiration time of the token expressed in seconds counting from the moment of receiving it.",
                        "type": "integer",
                        "format": "int64",
                        "example": 7200
                    },
                    "access_token": {
                        "title": "Access Token",
                        "description": "The Access Token (JWT). Valid for 8h.",
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                    },
                    "refresh_token": {
                        "title": "Refresh Access Token",
                        "description": "The standard Refresh Token. Valid for 24h.",
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserLogin"
                }
            },
...
            "UserLoginRequest": {
                "title": "Login User Request",
                "description": "Login User request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email Address",
                        "type": "string",
                        "example": "john@doe.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string",
                        "example": "passwordCreate12"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "title": "Create User Request",
                "description": "Create User request body data",
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "first_name": {
                        "title": "First name",
                        "description": "User's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "User's last name",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User's email address",
                        "type": "string",
                        "example": "john@doe.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password must be at least 8 characters long and contain a mix of letters and numbers.",
                        "type": "string",
                        "example": "passwordCreate12"
                    },
                    "password_confirmation": {
                        "title": "Password confirmation",
                        "description": "Password confirmation",
                        "type": "string",
                        "example": "passwordCreate12"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://gigs.devs/oauth/authorize",
                        "tokenUrl": "http://gigs.devs/oauth/token",
                        "refreshUrl": "http://gigs.devs/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}